Create writingMethods.java
Lab instructions
10 of 10 completed


inLabInstructionsPart~EjjO5VaAEe-SYAr_xnWqGQ
​
In this lab, you will repurpose your code from the driving simulation you have already written, focusing on breaking your code into different methods to isolate various functionalities within your program. Your primary goal is to maintain the basic functionalities of the original simulation, allowing users to control the engine, gear, and acceleration, with an option to exit from the simulation.

To achieve this, you will display a car dashboard showing the current speed, engine state (on/off), and gear (Park, Drive, Reverse). Additionally, you will provide a menu for users to interact with the car, enabling them to turn the engine on and off, change gears, accelerate the car (when the engine is on and not in Park), apply brakes (when the engine is on and speed is greater than 5), and exit the simulation. Through this hands-on experience, you will enhance your ability to structure your code more effectively by isolating different functionalities into distinct methods.

Goal

Repurpose and refactor the driving simulation code to isolate different functionalities into distinct methods, while maintaining the core features of controlling the engine, gear, acceleration, and providing an option to exit the simulation.

 🖥️  Note: When encountering this icon, it’s time to get into your IDE and start coding!  

In your lab environment, open IntelliJ by double-clicking on the icon