Create Coffee.java
In this task, you'll examine the class to understand how it represents coffee using attributes like name, roast, caffeine level, and price. You'll also explore how methods within the class set and utilize these attributes, providing insight into object-oriented programming in Java.

To begin, expand the src folder and open the provided Coffee class.

Observe that four attributes define Coffee:

name: a String representing, for example, “Espresso.”

roast: a String representing how strong the coffee must be, for example, “medium.”

caffeineLevelInMg: an int storing the caffeine level in milligrams, but its value is calculated based on the roast. For example, if a medium roast is selected, the caffeine level would be 50 mg.

price: a double representing the price of the coffee in dollars (for example, 3.45).

There is a parameterized constructor with three arguments. It assigns these values to the corresponding attributes and then calls the setCaffeineLevel() method. This ensures the caffeine level is set based on the chosen roast.

Remember

Methods can be called inside other methods. After all, you’ve been calling all your methods inside the main() method! 

A constructor is also a method, although it is slightly special! (to recall, you can refer to Constructors: Instantiating Objects with Different States)

The method setCaffeineLevel() checks the roast attribute value using an if-else if statement and assigns the appropriate value to the caffeineLevel attribute.

The grindBeans() and brewCoffee() methods simulate the coffee preparation, and the printInfo() method is defined to print the name, roast, and caffeineLevel of the coffee in a formatted way.


Complete