Create CoffeeMachine.java
Instructions
Lab instructions
9 of 10 completed


inLabInstructionsPart~1ZrJIVZ4Ee-a8xKK8ZhIJQ
​
It’s time to finally observe how the CoffeeMaker class utilizes polymorphism to handle different coffee types. You’ve already created the objects of the Espresso and Latte subclasses. However, before starting on the next coding steps mentioned below, you must complete an important task.

You must comment out the code under TODO 8, TODO 9, TODO 10, TODO 18, TODO 19, and TODO 20 because the grindBeans() and brewCoffee() methods are now being called by the prepareCoffee() method in the CoffeeMaker class. So, you do not need to call them individually for your Espresso and Latte objects. The printInfo() method also does not need to be called here.

Tip

How do you comment out code? 

To comment out a single line of code, you can simply type // (double slash) characters at the beginning of the line of code. For example, 

// myFavoriteEspresso.grindBeans();

The comments tell Java that this particular line in the code must not be executed.

Why not delete the code, you ask? Sometimes, you’d like to comment out code instead of deleting it so that you can verify that the rest of the code works well.

 🖥️  It’s time to get coding!  

Now, complete the following steps inside the main() method in the provided CoffeeMachine class:

🖥️ TODO 25: Create the object of CoffeeMaker class. For example: CoffeeMaker cafeCoffeeMaker = new CoffeeMaker();

This comment is at the beginning of the main() method after the Scanner class object has been created. The creation of the CoffeeMaker object is not condition-dependent, and it will be required inside both cases in the switch statement; therefore, you should create it outside the switch statement and the while loop.

🖥️ TODO 26: Call the prepareCoffee() method on your CoffeeMaker object and pass the Espresso object as an argument.

🖥️ TODO 27: Call the printInfo() method on your Espresso object using the dot operator.

🖥️ TODO 28:  Call the prepareCoffee() method on your CoffeeMaker object and pass the Latte object as an argument.

🖥️ TODO 29: Call the printInfo() method on your Latte object using the dot operator.

Remember

When you create Espresso and Latte objects, they are essentially Coffee objects due to inheritance.

When you pass an Espresso or Latte object to the prepareCoffee() method, the correct overridden grindBeans() and brewCoffee() methods are invoked based on the actual type of object at runtime.

Although the prepareCoffee() method is called twice for different coffee options in the switch case statement, it is actually executed only once based on your coffee selection and using the appropriate Coffee type object as the argument. This ensures proper implementation of polymorphism in the code.


Complete