Create Sidekick.java
nstructions
Lab instructions
6 of 6 completed


inLabInstructionsPart~JpM6V1jBEe-a8xKK8ZhIJQ
​
Toni needs your help to ensure that SideKick's properties are secure and validated. By following these steps, you'll set up a menu system for cleaning and cooking, validate property values, and understand the importance of access specifiers. Implementing validation is crucial for maintaining data integrity, enhancing security, improving user experience, and ensuring the robustness of the SideKick system. Proper validation prevents errors from invalid data, reduces security risks, helps users understand and correct mistakes, and makes the code easier to maintain and debug.


 🖥️  It’s time to get coding!  

🖥️ TODO 7: Go to the SideKick.java file and create the setter method in the same way as you did for the getter method. Ensure that you follow conventions and that the setter method is public. For more information on setter methods, refer to the video Encapsulation and data hiding in action.

🖥️ TODO 8: In the SideKick.java file, check the setter method you just created in TODO 7. If the value of the parameter being passed is not between 1 and 3, set it to 1. 

Why did you do this? Well, one important reason to create setter methods and not allow direct access is to ensure that the property's value is valid. Since the property's value can only be between 1 and 3, if an invalid value such as 200 is sent, you are preventing the value from representing an invalid value and setting it to a default value of 1. 	

TIP

A setter should always ensure that the value being assigned to a property is a valid value for the property, with reference to the context of the class or program.

12345678
     // mutator/setter method for modeOfOperation
    /** TODO 7: Implement a setter for "modeOfOperation" using Java convention   **/

    /** TODO 8: In the setter method you created in TODO 6 put in a condition
     *          to ensure that if the value being passed is not between 1 and 3
     *          then set it to the default of 1 - which represents cleaning
     **/

🖥️ TODO 9: Go to Main.java. Set the value of the property modeOfOperation to have a value of 3, using the setter method you created in TODO 7 and TODO 8.

🖥️ TODO 10: In Main.java, use the method takeAction() of the object sideKickObject, which you created in TODO 2.

1234567
        /** TODO 9: Set the value of the property "modeOfOperation" of the object
         *        "sideKickObject", using the setter you created in TODO 8, to a
         *        a value of 3
         **/

        /** TODO 10: Call the method takeAction() of the object "sideKickObject" **/

What did you observe? It executed case 3 and recharged the battery, right? Why? This is because you changed the value of modeOfOperation to 3 in TODO 9, using the public setter method.

12345678910111213141516171819
     // method to take action
     public void takeAction() {
        switch (modeOfOperation) {
            case 1:
                cleanHouse();
                System.out.println("SideKick cleaning completed.");
                break;
            case 2:
                System.out.println("SideKick cooking completed.");
                break;

🖥️ TODO 11: Return to SideKick.java and protect the method cleanHouse() from external access. How do you protect properties and methods from external access (such as from access from outside the class)? You can do this by using private in place of public. 

123456789101112131415
 /** TODO 11: The method cleanHouse() will only be used internally.
     *           Protect it with "private"
     **/
     // method to clean
    public void cleanHouse() {
        System.out.println("Get the vacuum cleaner.....");
        System.out.println("Put the dust bag in vacuum.....");
        System.out.println("Go to Living room and clean.....");
        System.out.println("Go to bedroom and clean.....");
        System.out.println("Go to kitchen and clean.....");

🖥️ TODO 12: Go to Main.java and use the setter method of the property modeOfOperation to pass the value 1 and set the value of modeOfOperation to 1. Then, call the method takeAction(). 

1234567
        /** TODO 12: Set the value of the property "modeOfOperation" of the object
         *          "sideKickObject to a value of 1
         *           using the setter method you created in TODO 7 and TODO 8.
         *           After that is done, call the method takeAction() using object
         *           "sideKickObject" and the dot operator.
         **/

Now, run the main program. 

Does it make a difference to the execution of the cleaning by SideKick? No, it does not. Remember the execution of a method remains the same, irrespective of the access specifier. It just changes the area where the property or method can be accessed. When you make it private it can be accessed only within the class, but it executes in the same way. 

Great work! You've successfully ensured that SideKick's properties are validated and secured by using getter and setter methods in combination with the private access specifier. This not only enhances the functionality of the robot but also maintains the integrity of its operations. Keep up the excellent coding!

Complete