Create Animal.java
Instructions
Lab instructions
7 of 9 completed


inLabInstructionsPart~vz8qE1V_Ee-WRxLcDuIdKQ
​
The base class of all the animals that we are going to create is the Animal class, and it has the following properties:

private String nameOfAnimal;

private int weight;

private int height;

private int age;

This class implements the Eat interface. The Eat interface has two methods:

eatingFood()

eatingCompleted()

The Animal class only implemented one of these methods, eatingFood(). As a result, it becomes abstract. Now, all child classes of the Animal class would have to implement the eatingCompleted() method, or they will also become abstract. The entire range of animals in the zoo will have the Animal class as its parent for now. 

 🖥️  It’s time to get coding!  

🖥️ TODO 1: Create a new child class Tiger that extends the Animal class named "Tiger".  

Create the following properties: number of stripes, speed, and sound level of roar.

Ensure that the variables' names represent the property they signify—for example, numberOfStripes for the number of stripes value. Also, ensure that all properties are protected using private.

Create the getter and setter methods for all the properties. Remember that you can right-click on Tiger.java and select generate to let IntelliJ automatically generate the getter and setter methods instead of coding them manually. Ensure that the getter and setter methods are public so that the properties are accessible indirectly from outside the class.

Ensure that the default constructor of the Tiger class calls the single-argument constructor of the parent class, passing the name as “Tiger” whenever you create an object using the default constructor of Tiger. This guarantees that the property nameOfAnimal is set to the type of animal, which is a tiger in this case, as soon as it is created and does not contain the value "Unknown Animal" set by the default constructor of the Animal class.

123456789
/** TODO 1: extend the class Animal to create a new
 * land based animal named "Tiger" which
 * extends Animal.
 * Extra properties of class "Tiger" are:
 * 1. number of stripes
 * 2. speed
 * 3. sound level of roar
 *
 **/
🖥️ TODO 2: Create a new child Dolphin class which extends the Animal class named "Dolphin". 

Create the following properties: color of dolphin and swimming speed.

Ensure that the variables' names represent the property they signify—for example, swimmingSpeed. Ensure all properties are protected using private.

Create the getter and setter methods for all the properties.  Ensure getter and setter methods are public so that the properties are accessible indirectly from outside the class. 

Ensure that the default constructor of the Dolphin class calls the single argument constructor of the parent class, passing the name as “Dolphin” whenever you create an object using the default constructor of Dolphin.

12345678
/** TODO 2: extend the class Animal to create a new
 * water based animal named "Dolphin" which
 * extends Animal
 * Extra properties of class "Dolphin" include:
 * 1. color of dolphin
 * 2. swimming speed
 *
 **/

Complete