Create Penguin.java
Instructions
Lab instructions
7 of 9 completed


inLabInstructionsPart~SNxIelWCEe-yWAr__tcdwQ
​
Now, try to create something special - a multifunctional class for a penguin that can walk and swim! Why is this important? By designing a class that can walk and swim, you'll demonstrate the power of inheritance and interfaces to handle complex behaviors. Imagine how versatile and efficient your code will become! This task challenges you to think creatively and apply what you've learned to make your code scalable and maintainable. 

 🖥️  It’s time to get coding!  

🖥️ TODO 9: Create a new child class of the Animal class named Penguin. 

Create a property isSwimming of boolean type to indicate whether the penguin is swimming or walking. Also, create properties walkSpeed and swimSpeed. Ensure all properties are protected using private.

Create the setters and getters of the properties. Ensure the getter and setter methods are public so that the properties are accessible indirectly from outside the class.

Ensure that when the default constructor of the Penguin is called, the name of the animal Penguin is passed to the constructor of the parent class Animal.

Override the eatingFood() method and implement the method eatingCompleted() of the Eat interface like the Dolphin class. Why? Because penguins eat fish like dolphins! 

Implement both the Walk and the Swim interfaces.

1
/** TODO 9: create a class "Penguin" from the "Animal" class **/
🖥️ TODO 10: Modify the menu system you set up in TODO  8 to choose a penguin. Remember, if you choose a penguin, you have to ask if it is walking or swimming before you can call the methods walking() or swimming(). Why do you think this is important? Because it cannot walk and swim at the same time!

Create a Penguin object after the Dolphin object. 

Modify the menu system in the method animalChoiceMenu()to include the option to select a penguin.

Add case 3 in the menu and show the same functionality for a penguin as for a tiger and dolphin. 

1
/** TODO 10: integrate the choice to pick a "penguin" in the menu system **/