Create SendButtonClass.java
Implementing interfaces
Introduction
An abstract method is useful for hiding detailed technical information. But what if you use more than one abstract method in your code? In these instances, you can use an interface as a collection of prototype declarations.

This reading will teach you how to employ interfaces in your applications.

Scenario
You’re building a messaging application for a social-media start-up company. The app requires three buttons: send a message, react to a message, and forward a message. You can build this application by implementing an interface. 

To begin, you’ll design an abstract button class that acts as a base for other buttons and an interface defining the methods the button classes must implement.

Interfaces overview
Let’s begin by exploring the term interface.

An object often displays the characteristics of more than one class. For example, a frog behaves like an animal on land, while it also has certain features of aquatic animals. In OOP terminology, this is referred to as multiple inheritance. 

Java doesn’t support the feature of multiple inheritance, as the compiler may not be able to distinguish between a method of the same name inherited by a subclass from multiple ancestors.

Interfaces are useful when the problem needs to be solved using multiple inheritance. It also lets you define a contract, which is a specification of a set of methods that a class must implement. 

An interface in Java is a collection of abstract methods that a class must implement. Since the methods are abstract, you need to provide only their method signatures (like names, parameters, and return types) without any implementation (in other words, there’s no method body inside curly braces). An interface defines what an implementing class should do without specifying how it should do it.

Java uses the interface keyword to declare an interface and the implements keyword in the declaration of a subclass.

Remember

For more information, refer to Module 3 of the Amazon Junior Software Developer Certificate course. 

Create the abstract class
In the scenario described above, the app should have buttons for different purposes, but all have a caption as the common property. As a first step for developing the social media messaging app, declare the abstract base class.

The setCaption() and getCaption() methods are non-abstract (concrete), so they are reused in the subclasses without overriding.

Remember

For more information, refer to Module 3 of the Amazon Junior Software Developer Certificate course.

1234567891011
abstract class AbstractButtonClass {
    String caption;

    public void setCaption(String caption) {
        this.caption = caption;
    }

    public String getCaption() {
        return this.caption;
    }

Create the interface
The app requires buttons for different actions, like Send, Forward, and Like. All of these buttons must be subclasses of the AbstractButtonClass, and each must perform a relevant action when pressed. 

To create these buttons, you must first declare a ButtonPress interface using the buttonPress() method, which is an abstract method (In other words, it doesn’t have any method body).

123
interface ButtonPress {
    void buttonPress();
}
Build the SendButtonClass
The next step is to create classes for each button required in the social media messaging app. You can do this by extending the AbstractButtonClass. While each subclass may have additional attributes and methods, the ButtonPress interface and the buttonPress() method must be implemented.

The SendButtonClass also requires additional attributes:

sendTo (such as email ID of the recipient) 

and message (the text of the message to be sent). 

It also defines the setSendTo() and setMessage() methods. Importantly, it also defines the buttonPress() method, which is necessary to implement the interface.