Create Date.java
Instructions
Lab instructions
9 of 9 completed


inLabInstructionsPart~S86LyWPBEe-Pwg6Ey4MMBQ
​
You’re off to a great start! The next step is to create the Date class, as an attribute of the Person class is a Date object (indicating a composition relationship). You’ll also need to include the getDate() method, which returns the string representation of the date.

 🖥️  It’s time to get coding!  

 🖥️ TODO 2: First, declare the Date class with the integer attributes date, month, and year. 

1234567891011
class Date {
    int dd;
    int mm;
    int yy;

    Date(int d, int m, int y) {
        this.dd = d;
        this.mm = m;
        this.yy = y;
    }

 🖥️ TODO 3: Then, include the getDate() method, which returns the string representation of the date.

1234
String getDate() {
    return this.dd + "-" + this.mm + "-" + this.yy;
}


Complete