Create EmployeeInterface.java
Instructions
Lab instructions
12 of 12 completed


inLabInstructionsPart~OE--8E9rEe-u-BJPdixGZQ
​
Now that you've laid the foundation with the Person class, it's time to define specific behaviors for employees. But before you implement the EmployeeInterface, let's take a moment to think about the role of interfaces in our system. Consider the various actions an employee might perform in a college setting. What responsibilities are unique to employees, as opposed to students or general persons? How can you use an interface to define a contract for these employee-specific behaviors without tying ourselves to a specific implementation? Think about how this separation of interface and implementation might make our system more flexible and easier to maintain in the long run.

Think about what kind of abstract methods the interface needs to compute tax and salary. Open the starter code for EmployeeInterface from the src folder. Each employee draws a salary and pays tax. You must create an EmployeeInterface with computeSalary() and computeTax() methods to track these actions as follows:

🖥️  It’s time to get coding!

🖥️ TODO 3: Add computeSalary() and computeTax() methods. Both have double as their return type. Remember, the methods in the interface do not have a method body.

123
interface EmployeeInterface {
    //TODO 3: add computeSalary() and computeTax() methods
}

Complete