Create Interface with Concrete methods.java
Interface with Concrete methods
So far, you’ve just explored using the abstract methods in an interface. From Java 8 onwards, the interface can now have concrete methods (methods that have functionality). 

Before version 8, if you wanted to add a new method to the interface, you would have had to provide its implementation code in the same class. As a result, any change in the interface by adding a method would also affect the implementation class.

With Java version 8, an interface can have a concrete method declared with a default keyword. The implementing class does not need to override it but may do so if needed.

So, how does this work in practice? 

Over at ConnectSphere, the company’s MediaPlayer interface has some abstract and default methods. These are forward() and backward(), which skip the playing position.

12345678910111213
interface ActionController {
    void start();
    void pause();
    void stop();

    default void forward() {
        System.out.println("skip forward by 10 sec");
    }

    default void backward() {

The MediaPlayer class can now implement the ActionController. While it needs to override the abstract methods in the ActionController (start, pause, and stop), it doesn't necessarily override the concrete methods with default keywords (forward and backward).

123456789101112131415
class MediaPlayer implements ActionController {

    public void start() {
        System.out.println("Player has started");
    }

    public void pause() {
        System.out.println("player paused");
    }


Press the Run button on the following interactive code block to test the behavior of the concrete methods in an interface.