Create Comparable interface.java
Comparable interface
Java’s standard library has several built-in interfaces, including Serialzable, Clonable, and Comparable interfaces.

The Comparable interface defines how objects are compared. It is generally implemented to sort an array of objects. It has a compareTo() method, which needs to be implemented by a concrete class.

The following interactive code uses the Comparable interface to compare two objects of the Length class. The length class has two attributes: cm and mm. The compareTo() method converts the cm value to mm for two objects and returns 1 or 0, depending on whose mm value is greater or less.