Create Course.java
Instructions
Lab instructions
15 of 15 completed


inLabInstructionsPart~hYbzLwQlEfC1RhKc0IDfxw
​
You’re off to a great start! The next step is to create the Course class. First, open the code for the course class, which is declared abstract, from the src folder.

🖥️ It’s time to get coding!

🖥️ TODO 3: Add the instance variables subject, instructor (String) and the course fee (int), assignmentMarks and quizMarks (int).

The variable subject is an object of the Subject class, define a Constructor for Course class.

123456789
abstract class Course {
    //TODO 3: include instance variables for Course class and
 complete constructor code

    Course(Subject subject, String instructor, int fee) {
        
    }
}


Complete