Create FinallyTestOne.java
When no Exception occurs
Run this block of code and observe the output.

1234567891011121314
public class FinallyTestOne {
    public static void main(String[] args) {
        try {
            int[] array = {10, 20, 30, 40, 50};
            System.out.println(array[4]);
        } catch(ArrayIndexOutOfBoundsException e) {
            System.out.println(e.getMessage());
        } finally {
            System.out.println("In the finally block!");
        }

        System.out.println("More code");
    }
}
Reset
50
In the finally block!
More code

The code does not cause an Exception because it attempts to print the element at index 4 in the array.  As a result 50 is printed and the catch block handling the ArrayIndexOutOfBoundsException is not invoked.  Instead, the finally block is executed printing the message: “In the finally block!” After that “More code” is printed, indicating that any code following the try-catch -finally block is executed smoothl