Create FinallyTestTwo.java
But what if an Exception occurs and there is a catch block to handle it? 

Run the following code to explore this scenario:

123456789101112131415
public class FinallyTestTwo {
    public static void main(String[] args) {
        try {
            int[] array = {10, 20, 30, 40, 50};
            System.out.println(array[5]);
        } catch(ArrayIndexOutOfBoundsException e) {
            System.out.println(e.getMessage());
        } finally {
            System.out.println("In the finally block!");
        }

        System.out.println("More code");

    }
}
Reset
5
In the finally block!
More code

This code causes the ArrayIndexOutOfBoundsException. As a result, the catch block prints out the Exception message. Then, the code executes the finally block before moving on to the line of code placed after this try-catch -finally block, and “More code” is printed to indicate the smooth execution of the rest of the program.