Create FinallyTestThree.java
The final scenario was if an Exception occurs and there is no catch block to handle it.

Press the Run button to execute the following code and observe what happens:

1234567891011121314
public class FinallyTestThree {
    public static void main(String[] args) {
        try {
            int[] array = {10, 20, 30, 40, 50};
            System.out.println(array[5]);
        } catch(ArithmeticException e) {
            System.out.println(e.getMessage());
        } finally {
            System.out.println("In the finally block!");
        }

        System.out.println("More code");
    }
}
Reset
Exception in thread "main" java.lang.ArrayIndexOutOfBoundsException: 5


This code throws an ArrayIndexOutOfBoundsException. However, the catch block is designed to handle the ArithmeticException. Since no code is present to handle the ArrayIndexOutOfBoundsException, the finally block is executed printing the message: “In the finally block!”, then the exception details is printed, and because the Exception is not handled, the program is abnormally terminated, and the line of code placed after the try-catch -finallyblock never gets executed.