Create FinallyTestFour.java
More examples of using the finally block
Another interesting thing about the finally block is that it still gets executed even if the enclosing method returns.

For example, the following code block has a method named containsAnE(String word), which returns a String by performing a boolean check. The return keyword is present in the try block. A return keyword is also present in the catch block to return in case an Exception occurs.

In the main() method, If you call the containsAnE method with the word “Hello”, it causes no Exception. The method returns from the try block. However, the message: “In the finally block!” would still print. 

If you call the containsAnE method with null passed in for the word, it will cause a NullPointerException. The catch block is invoked and handles the Exception. Though there is a return statement in the catch block, the finally block gets executed before control is handed over to the calling method, and the next line of code gets executed.

You can run this code to find out for yourself.

12345678910111213141516171819
public class FinallyTestFour {

    private static String containsAnE(String word) {
        try {
            return word.contains("E") ? "Contain an E" : "No E found!";
        } catch(NullPointerException e) {
            System.out.println("A NullPointerException occurred!");
            return "Return from catch";
        } finally {
            System.out.println("In the finally block!");
        }
    }

    public static void main(String[] args) {
	   //containsAnE("Hello");
        containsAnE(null);
        System.out.println("More code");
    }
}
Reset
In the finally block!
More code

This behavior is specifically helpful to avoid having clean up code accidentally bypassed due to a return statement in your code. In a more practical scenario, you’ll place clean up code within your finally block rather than simply printing out a message.

Conclusion
Now that you have a clear knowledge of the finally block and its role in cleaning up resources or performing essential tasks, you will find it useful to add it to your program. While you focus on discovering why the code to access information in the database is having issues, a finally block will ensure that the connection to the database is closed when the program terminates. By putting clean-up code in the finally block you will be following best practice. 

Are you excited by the possibilities of the finally block? Later versions of Java have introduced try-with-resources, an alternative to using the finally block and is explored in the next video.