Create CreateFile.java
Using the File class to create a new file
To create a new file using the File class, you must first create a File class object. This object will take the file name as an argument in its constructor. If you want to keep the file in a particular directory, you can add the path to the file name at the beginning  and append the file extension at the end:

"src/" + fileName + ".txt";

Before running this code, you'll need to import the following necessary classes.