Create bufferedWriter.java
Using the BufferedWriter class to write data to a file
The BufferedWriter class also provides overloaded versions of write() methods, similar to the FileOutputStream class. Fortunately, it works with character arrays and String values directly. Like earlier, creating a BufferedWriter object also requires the creation of a FileWriter object (FileWriter is a convenience class designed for writing streams of characters). 

Tip

If you need to append data to an existing file, both the FileOutputStream and FileWriter classes provide overloaded constructors that accept a second boolean argument. If this argument is ‘true’, data will be appended to the end of the file instead of overwriting it.

FileOutputStream(File file, boolean append)

FileWriter(File file, boolean append)

These constructors are helpful when you want to add new data to an existing file without losing the previous content. In upcoming lessons, you'll have the opportunity to practice and learn how to use these "append" methods to write to files without overwriting the existing data.