Create LibraryManagementSystem.java
Coursera
Amazon
Introduction to Software Development
Module 1
Working with Strings
Working with Strings
In this lab, you’ll use various methods to manipulate Strings in Java. You will check String length, shorten Strings, and search for specific parts of a String.

Let’s use Riley as an example. Riley has a problem. She is developing a library management system for clients to better organize and catalog their book inventory. However, there is an issue with capturing book titles: when they are too long, they do not display correctly in the management system.

To ensure book titles display correctly, they must be no more than 20 characters long. Riley needs to check the title length and shorten it if it exceeds this limit. Additionally, Riley wants the system to search for specific words in a book title to enhance functionality.

You’ve been provided with the starter code. To help Riley, you must write the necessary code to accomplish the requested tasks. Let’s get started! 

Reflection point: Understanding arrays
Previously, you learned that an array in Java is a collection of elements of the same type stored in contiguous memory locations. Arrays allow efficient data access and manipulation using an index and help handle multiple values with a single variable. Therefore, a String is a sequence of characters represented internally as an array of the char data type.

Understanding arrays is crucial because it forms the basis of how Strings and other data structures work in Java. By grasping the concept of arrays, you will be better equipped to manipulate data efficiently, as you will be doing in this lab.

Goal

Can you help Riley by checking book title lengths, shortening long titles, and searching for specific words in book titles?

 🖥️  Note: When encountering this icon, it’s time to get into your IDE and start coding!