Update and rename Course-01-Introduction-to-Software-Development/Module-02: Control flow :Statements and loops/Concepts.md to Course-01-Introduction-to-Software-Development/Module-02: Control flow :Statements and loops/Lab/Building a schedule/Main.java
Coursera
Amazon
Introduction to Software Development
Module 2
Building a schedule
Building a schedule
In this lab, you’ll create and access a schedule by using a switch statement. switch statements provide a structured and efficient way to define, access, and manage schedules in code, promoting clarity, maintainability, and scalability.

Let’s use Toni as an example. He has a lot of work to catch up on and keeps forgetting what needs to be done, especially after coming back from his road trip. 

Goal

Can you help Toni by creating a program that displays the schedule for the day when the day of the week is given in numeric form? For example, you could represent Monday as 1, Tuesday as 2, and so on.

To help Toni, you’ll need to write the necessary code to accomplish the tasks requested. Let’s get started! 

 🖥️  Note: When encountering this icon, it’s time to get into your IDE and start coding!  

In your lab environment, open IntelliJ by double-clicking on the icon.

""
When the IntelliJ IDE opens, you will be brought to the starter code.