Create TextEditorUndo.java
Coursera
Instructions
Lab instructions
11 of 11 completed


inLabInstructionsPart~j-pm22rQEe-Pwg6Ey4MMBQ
​
Open the starter code for com.lesson.lab from the src folder, in which you’ll find the following classes:

TextEditorUndo Class

This class includes attributes and methods to manage the stack operations for an undo feature in a text editor.

Attributes:

String[] stack: An array of strings for storing actions.

int top: An index to track the top of the stack, indicating the most recent action.

Methods: Includes methods to push, pop, peek, and display actions in the stack.

Main Class

This class contains the code to execute the program, handle exceptions, and interact with the TextEditorUndo class.

🖥️ It’s time to get coding!

To begin, use the predefined class TextEditorUndo to initialize the stack array, as the constructor has already been created.

TODO 1: initialize the stack array with the size (e.g. 10).

TODO 2: set the top variable to -1 to indicate that the stack is empty. Variable top, which is already defined in the TextEditorUndo class, represents the index of the top element in the stack.

1234567
public class TextEditorUndo{
    public TextEditorUndo(int size) {
            // TODO 1: initialize the stack array
            // TODO 2: set variable top index to -1 initially
        }
}


Complete